Move FieldNames to its own module
